refactor: 도메인 관련 클래스 및 테스트 패키지 구조 재정비
