+++ src/main/java/janggi/controller/Controller.java
@@ -1,11 +1,11 @@
 package janggi.controller;
 
-import janggi.domain.Board;
-import janggi.domain.BoardFactory;
+import janggi.domain.board.Board;
+import janggi.domain.board.BoardFactory;
 import janggi.domain.position.Column;
 import janggi.domain.position.Position;
 import janggi.domain.position.Row;
-import janggi.domain.Team;
+import janggi.domain.piece.Team;
 import janggi.dto.BoardDto;
 import janggi.dto.PositionInputDto;
 import janggi.dto.TeamInputDto;
