+++ src/main/java/janggi/domain/board/BoardFactory.java
@@ -1,5 +1,8 @@
-package janggi.domain;
+package janggi.domain.board;
 
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.PieceType;
+import janggi.domain.piece.Team;
 import janggi.domain.position.Column;
 import janggi.domain.position.Position;
 import janggi.domain.position.Row;
