+++ src/test/java/janggi/domain/movestrategy/MaMoveStrategyTest.java
@@ -1,8 +1,8 @@
 package janggi.domain.movestrategy;
 
-import janggi.domain.BoardState;
+import janggi.domain.board.BoardState;
 import janggi.domain.position.Column;
-import janggi.domain.Piece;
+import janggi.domain.piece.Piece;
 import janggi.domain.position.Position;
 import janggi.domain.position.Row;
 import org.junit.jupiter.api.Test;
