+++ src/test/java/janggi/domain/movestrategy/PoMoveStrategyTest.java
@@ -1,12 +1,12 @@
 package janggi.domain.movestrategy;
 
-import janggi.domain.BoardState;
+import janggi.domain.board.BoardState;
 import janggi.domain.position.Column;
-import janggi.domain.Piece;
-import janggi.domain.PieceType;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.PieceType;
 import janggi.domain.position.Position;
 import janggi.domain.position.Row;
-import janggi.domain.Team;
+import janggi.domain.piece.Team;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
