+++ src/test/java/janggi/domain/movestrategy/board/BoardFactoryTest.java
@@ -1,11 +1,15 @@
-package janggi.domain;
+package janggi.domain.movestrategy.board;
 
+import janggi.domain.board.Board;
+import janggi.domain.board.BoardFactory;
 import janggi.domain.movestrategy.ChaMoveStrategy;
 import janggi.domain.movestrategy.GungseongBoundMoveStrategy;
 import janggi.domain.movestrategy.JolMoveStrategy;
 import janggi.domain.movestrategy.MaMoveStrategy;
 import janggi.domain.movestrategy.PoMoveStrategy;
 import janggi.domain.movestrategy.SangMoveStrategy;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Team;
 import janggi.domain.position.Column;
 import janggi.domain.position.Position;
 import janggi.domain.position.Row;
